EveSafe
“EveSafe” is a fully loaded android application designed to ensure
safety, especially women.  During a dread situation when you are struck
in panic; nevertheless, it is imperative that your  loved ones are
noticed, and that the police authorities are able to pin down your exact
location. So that you  can  be saved  from  that  situation. And this
fender application will act as your safety companion to guide you  in
those dread  situation